backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
redis==5.0.8
pandas==2.2.3
numpy==2.1.1
google-generativeai==0.8.3
apscheduler==3.10.4
httpx==0.27.2
python-multipart==0.0.12
alembic==1.13.2